requirements.txt
asgiref==3.5.2
django==3.2.15
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.3.0
isort==5.10.1
psycopg2==2.9.3
python-dotenv==0.20.0
Pillow==9.2.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==22.1
drf-extra-fields==3.4.0
reportlab==3.6.11